refactor: remove old reporting code
