+++ src/brisk/reporting/components/navbar.html
