+++ src/brisk/reporting/components/plot.html
