+++ src/brisk/reporting/components/table.html
