+++ src/brisk/reporting/js/core/app.js
