+++ src/brisk/reporting/js/renderers/dataset_page.js
