+++ src/brisk/reporting/js/renderers/experiment_group_card.js
