+++ src/brisk/reporting/js/renderers/home.js
