+++ src/brisk/reporting/js/renderers/plot.js
