+++ src/brisk/reporting/js/renderers/table.js
