--- src/brisk/reporting/styles/index.css
@@ -1,218 +1,0 @@
-/* Styling the top bar */
-header {
-    background-color: #1B1B1B;
-    color: #E0E0E0;
-    text-align: left;
-    padding: 20px 0;
-    padding-left: 10px;
-    font-size: 28px;
-    width: 100vw;
-}
-
-body {
-    background-color: #333333;
-    font-family: 'Montserrat', sans-serif; /* Global font setting */
-    font-weight: 400;
-    font-size: 16px; /* Set base font size */
-    line-height: 1.6; /* Adjust line height for better readability */
-    margin: 0;
-    padding: 0;
-}
-
-/* Section styling for datasets */
-section {
-    padding: 10px;
-    margin: 10px 0;
-    /* margin-left: 20px; */
-    width: 45vw
-}
-
-/* Container for the sections */
-.section-container {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 20px;
-    padding: 20px;
-}
-
-/* Style for individual sections */
-.dataset-section {
-    padding: 10px;
-    color: #E0E0E0; 
-    border-radius: 8px;
-}
-
-/* Color backgrounds for dataset sections */
-.dataset-section:nth-child(odd) {
-    background-color: #14274A; 
-}
-
-.dataset-section:nth-child(even) {
-    background-color: #1E3A8A; 
-}
-
-/* Link styles */
-a {
-    text-decoration: none;
-    color: #E0E0E0;
-}
-
-a:hover {
-    text-decoration: underline;
-    color: #E0CA3C;
-}
-
-p {
-    font-size: 16px; /* Increase font size for paragraphs */
-    line-height: 1.6; /* Set line height to make text more readable */
-    color: #E0E0E0; /* Keep text color light gray for contrast */
-    margin: 15px 0; /* Add margin for spacing between paragraphs */
-}
-
-h1, h2, h3 {
-    font-weight: 700; /* Make headers bold */
-    line-height: 1.3; /* Tighter line height for headers */
-}
-
-h2 {
-    margin: 30px 0 15px 0; 
-    padding: 0; 
-    color: #E0E0E0; 
-    line-height: 1.3; 
-    margin-left: 10px;
-}
-
-ul {
-    list-style-type: none;
-    padding-left: 20px;
-    counter-reset: link-counter; /* Reset counter */
-}
-
-li {
-    counter-increment: link-counter;  /* Increment counter */
-}
-
-li::before {
-    content: counter(link-counter) ". ";  /* Display the counter */
-    color: #E0E0E0;
-}
-
-/* For table headers */
-th {
-    background-color: #14274A; 
-    color: #E0E0E0;
-    border: 2px solid #444444; 
-    padding: 12px;
-    font-weight: 600;
-}
-
-/* For table body */
-td {
-    background-color: #1E3A8A;
-    color: #E0E0E0; 
-    border: 1px solid #444444;
-    padding: 12px;
-    font-weight: 400;
-}
-
-/* Adding horizontal and vertical spacing */
-table {
-    border-collapse: collapse;
-    border-spacing: 0 8px;
-    width: 80%;
-    margin: 0 auto;
-}
-
-.experiment-section {
-    border-radius: 8px;
-    margin: 0;
-    padding: 20px;
-}
-
-/* First column */
-.experiment-section:nth-child(4n + 1) {
-    background-color: #14274A;
-}
-
-.experiment-section:nth-child(4n + 3) {
-    background-color: #1E3A8A;
-}
-
-/* Second column */
-.experiment-section:nth-child(4n + 2) {
-    background-color: #1E3A8A;
-}
-
-.experiment-section:nth-child(4n + 4) {
-    background-color: #14274A;
-}
-
-.section-header {
-    border-bottom: 2px solid #eee;
-    margin-bottom: 15px;
-    padding-bottom: 10px;
-}
-
-.section-header h2 {
-    color: #ffffff;
-    margin: 0;
-    font-size: 1.5em;
-}
-
-.section-header h3 {
-    color: #7f8c8d;
-    margin: 5px 0 0 0;
-    font-size: 1.2em;
-}
-
-.experiment-list {
-    padding: 10px 0;
-}
-
-.experiment-list h4 {
-    color: #ffffff;
-    margin: 0 0 10px 0;
-}
-
-.experiment-list ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0 0 15px 0;
-}
-
-.experiment-list li {
-    margin: 5px 0;
-}
-
-.dataset-link {
-    margin-top: 15px;
-    padding-top: 15px;
-    border-top: 1px solid #eee;
-}
-
-.dataset-link a {
-    color: #3498db;
-    text-decoration: none;
-}
-
-.dataset-link a:hover {
-    text-decoration: underline;
-}
-
-.sortable th {
-    cursor: pointer;
-    position: relative;
-    padding-right: 20px !important;
-}
-.sortable th::after {
-    content: '↕';
-    position: absolute;
-    right: 8px;
-    color: #666;
-}
-.sortable th.asc::after {
-    content: '↑';
-}
-.sortable th.desc::after {
-    content: '↓';
-}
