+++ src/brisk/reporting/styles/navbar.css
