+++ src/brisk/reporting/styles/table.css
