--- src/brisk/reporting/templates/dataset.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dataset Analysis - {{ group_name }}/{{ dataset_name }}</title>
-    <link rel="stylesheet" href="./dataset.css">
-    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700;800;900&display=swap" rel="stylesheet">
-</head>
-<body>
-    <header>
-        Dataset Analysis: {{ group_name }} / {{ dataset_name }}
-        <div style="font-size: 14px; color: grey;">
-            Version: {{ version }}
-        </div>
-    </header>
-    <div class="content">
-        {{ content | safe }}
-    </div>
-</body>
-</html>
